+++ http4k-contract/openapi/src/main/kotlin/org/http4k/contract/openapi/v3/model.kt
@@ -10,6 +10,7 @@ import org.http4k.lens.ParamMeta.ArrayParam
 import org.http4k.lens.ParamMeta.FileParam
 import org.http4k.lens.ParamMeta.ObjectParam
 
+@ConsistentCopyVisibility
 data class Api<NODE> internal constructor(
     val info: ApiInfo,
     val tags: List<Tag>,